requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
idpay==0.0.1
psycopg2==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
tzdata==2023.3
urllib3==2.0.4